Home: poll learn mount readiness instead of trusting boot snapshot
The boot-time config snapshot's learn_mount_ready is fetched before the
server's background automount finishes, so gating the Learn doorway on
it meant the card essentially never appeared. Extract the Sidebar's
bounded /api/config re-poll into a shared useLearnMountReady hook and
use it in both Sidebar and Home.
